requirements.txt
streamlit
pandas
numpy
pydeck
matplotlib
reportlab
arabic-reshaper
python-bidi
openpyxl
